Add example of creating and using lib in sysroot
Two short recipies with files along them were created. One of them
(mylib_%.bb) develops a static library and installs it in sysroot. It is
then available for the gcc-mechine-specific cross-compilation tool.

The hello-world_%.bb bitbake recipe uses the previously created mylib
static library. It is able to find the header file in the /use/include
directory (in sysroot) and the link all files against it.